fix bug (minify _layout.svelte)
